+++ CHANGELOG.md
@@ -6,11 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/airbnb/epoxy-ios/compare/0.1.0...HEAD)
 
-### Added
-- Added `SelectableStackView` for managing a stack of bar models with selection.
-
 ### Changed
 - Updated public let properties of public structs with memberwise initializers to be public var.
+- `BarStackView` now handles selection of bar models.
 
 ## [0.1.0](https://github.com/airbnb/epoxy-ios/compare/171f63da...0.1.0) - 2021-02-01